--- src/controllers/hashtags.controllers.js
@@ -1,17 +1,0 @@
-import { listTrendingHashtags } from "../repositories/hashtags.repository.js";
-
-const getHashtags = async (req, res) => {
-
-  try {
-
-    const trendingHashtags = await listTrendingHashtags();
-
-    res.send(trendingHashtags.rows);
-  } catch (error) {
-    console.log(error);
-    res.sendStatus(500);
-  }
-
-};
-
-export { getHashtags };
