+++ src/server.js
@@ -1,8 +1,8 @@
 import express from "express";
 import cors from "cors";
 import dotenv from "dotenv";
-import hashtagsRouters from "./routers/hashtags.router.js";
-import authRoutes from "../src/routes/authRoutes.js";
+import hashtagsRouters from "./routes/hashtagsRoutes.js";
+import authRoutes from "./routes/authRoutes.js";
 dotenv.config();
 
 const app = express();
